Create gvCaseInfo.vm file
File contains 4 parts - the actual global variable you need to create using the "EvaluateVelocityTemplate" option, followed by 3 local variables - ARC, FAIL, and lvCategory.  Keep in mind that this is just an example - you can add/remove/change things up however you need, just keep in mind that YOU NEED TO AGGREGATE on the global variable (and fields that will populate the global variable) if you want to see events in a case.  

Currently, this variable sits on a tier 2 ESM, so events that it sees are already correlated/aggregated.  If I need additional information from a Tier 1 ESM, I aggregate on that field, and that data will automatically be pulled into the case information using the global variable.